Move MultipleDeclarationsInspectionTests for QuickFixes into relevant files under RD.QuickFixes
