+++ RubberduckTests/Inspections/MultipleDeclarationsInspectionTests.cs
@@ -2,7 +2,6 @@ using System.Linq;
 using System.Threading;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Rubberduck.Inspections.Concrete;
-using Rubberduck.Inspections.QuickFixes;
 using Rubberduck.Parsing.Inspections.Resources;
 using Rubberduck.VBEditor.SafeComWrappers.Abstract;
 using RubberduckTests.Mocks;
@@ -133,117 +132,6 @@ End Sub";
             var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
 
             Assert.IsFalse(inspectionResults.Any());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void MultipleDeclarations_QuickFixWorks_Variables()
-        {
-            const string inputCode =
-@"Public Sub Foo()
-    Dim var1 As Integer, var2 As String
-End Sub";
-
-            const string expectedCode =
-@"Public Sub Foo()
-    Dim var1 As Integer
-Dim var2 As String
-
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new MultipleDeclarationsInspection(state);
-            var inspector = InspectionsHelper.GetInspector(inspection);
-            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
-
-            new SplitMultipleDeclarationsQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void MultipleDeclarations_QuickFixWorks_Constants()
-        {
-            const string inputCode =
-@"Public Sub Foo()
-    Const var1 As Integer = 9, var2 As String = ""test""
-End Sub";
-
-            const string expectedCode =
-@"Public Sub Foo()
-    Const var1 As Integer = 9
-Const var2 As String = ""test""
-
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new MultipleDeclarationsInspection(state);
-            var inspector = InspectionsHelper.GetInspector(inspection);
-            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
-
-            new SplitMultipleDeclarationsQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void MultipleDeclarations_QuickFixWorks_StaticVariables()
-        {
-            const string inputCode =
-@"Public Sub Foo()
-    Static var1 As Integer, var2 As String
-End Sub";
-
-            const string expectedCode =
-@"Public Sub Foo()
-    Static var1 As Integer
-Static var2 As String
-
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new MultipleDeclarationsInspection(state);
-            var inspector = InspectionsHelper.GetInspector(inspection);
-            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
-
-            new SplitMultipleDeclarationsQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void MultipleDeclarations_IgnoreQuickFixWorks()
-        {
-            const string inputCode =
-@"Public Sub Foo()
-    Dim var1 As Integer, var2 As String
-End Sub";
-
-            const string expectedCode =
-@"Public Sub Foo()
-'@Ignore MultipleDeclarations
-    Dim var1 As Integer, var2 As String
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new MultipleDeclarationsInspection(state);
-            var inspector = InspectionsHelper.GetInspector(inspection);
-            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
-
-            new IgnoreOnceQuickFix(state, new[] {inspection}).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
         }
 
         [TestMethod]
